// src/types/project.ts
export interface TaskResponse {
  id: number;
  project_id: number;
  task_action_id: number;
  task_name: string;
  complete_date: string;
  comment: string;
  email_alert: string;
  assigned_to: number;
  assigned_by: number;
  assigned_at: string;
  status: string;
}

export interface DocumentResponse {
  id: number;
  project_id: number;
  title: string;
  notes: string;
  date: string;
  filename: string;
}

export interface DBProject {
  id: string;
  project_name: string;
  status: string;
  user_id: string;
  address: string;
  city: string;
  zip: any;
  start_date: string;
  state: string;
  tasks: TaskResponse[];
  project_date: ProjectDateResponse[];
  documents: DocumentResponse[];
  project_contract: ProjectContractResponse;
  created_at: string;
  project_type?: { id: number, project_type: string, description?: string }
}